reset to size_t

+++ klib/klib.h
@@ -16,7 +16,7 @@ extern "C" {
 //   memcpy, memset, strcpy, strlen, itoa, atoi, ...
 //   kalloc, kree
 
-int strlen(const char *s);
+size_t strlen(const char *s);
 
 #ifdef __cplusplus
 }
